# services/api/openapi.yml
openapi: 3.0.0
info:
  description: Api endpoints for http://oddson.dev/
  version: "1.0.0"
  title: Odds on
paths:
  /:
    get:
      responses:
        404:
          description: "Not Found"
      x-amazon-apigateway-integration:
        uri: ${api_function}
        type: "aws_proxy"
        httpMethod: "POST"
  /new:
    post:
      summary: Creates a new challenge
      operationId: createChallenge
      responses:
        "200":
          description: Challenge created
          content:
            application/json:
              schema:
                type: object
                $ref: "#/components/schemas/Challenge"
        "400":
          description: "invalid input, object invalid"
        "409":
          description: an existing item already exists
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChallengeRequest"
      x-amazon-apigateway-integration:
        uri: ${api_function}
        type: "aws_proxy"
        httpMethod: "POST"
    options:
      responses:
        200:
          description: "Response to CORS"
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Headers:
              schema:
                type: string
          content: {}
      x-amazon-apigateway-integration:
        uri: ${api_function}
        type: "aws_proxy"
        httpMethod: "POST"

  /activate/{id}:
    post:
      summary: Victim activates the challenge by entering their odds
      operationId: activateChallenge
      responses:
        "200":
          description: Challenge activated
          content:
            application/json:
              schema:
                type: object
                $ref: "#/components/schemas/ChallengeResponse"
        "400":
          description: "invalid input, object invalid"
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: challenge id
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivateChallengeRequest"
      x-amazon-apigateway-integration:
        uri: ${api_function}
        type: "aws_proxy"
        httpMethod: "POST"
    options:
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: challenge id
      responses:
        200:
          description: "Response to CORS"
          content: {}
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Headers:
              schema:
                type: string
      x-amazon-apigateway-integration:
        uri: ${api_function}
        type: "aws_proxy"
        httpMethod: "POST"

  "/complete/{id}":
    post:
      summary: Complete the challenge by guessing the victim's odds
      operationId: completeChallenge
      responses:
        "200":
          description: Challenge outcome
          content:
            application/json:
              schema:
                type: object
                $ref: "#/components/schemas/CompleteChallengeResponse"
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: challenge id
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteChallengeRequest"
      x-amazon-apigateway-integration:
        uri: ${api_function}
        type: "aws_proxy"
        httpMethod: "POST"
    options:
      responses:
        200:
          description: "Response to CORS"
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Headers:
              schema:
                type: string
          content: {}
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: challenge id
      x-amazon-apigateway-integration:
        uri: ${api_function}
        type: "aws_proxy"
        httpMethod: "POST"

  /check/{id}:
    get:
      summary: Check the status of an existing challenge
      operationId: checkChallenge
      responses:
        "200":
          description: Challenge status
          content:
            application/json:
              schema:
                type: object
                $ref: "#/components/schemas/CheckResponse"
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: challenge id
      x-amazon-apigateway-integration:
        uri: ${api_function}
        type: "aws_proxy"
        httpMethod: "POST"
    options:
      responses:
        200:
          description: "Response to CORS"
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Headers:
              schema:
                type: string
          content: {}
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: challenge id
      x-amazon-apigateway-integration:
        uri: ${api_function}
        type: "aws_proxy"
        httpMethod: "POST"

components:
  x-amazon-apigateway-integrations:
    oddson_express:
      uri: ${api_function}
      type: "aws_proxy"
      httpMethod: "POST"

  schemas:
    # Response Models
    CheckResponse:
      allOf:
        - $ref: "#/components/schemas/Challenge"
      example:
        id: fe5d746a-e582-4cec-b2c8-d0c44e9108d1
        challenge: Do the thing
        odds: 20
        challenger:
          name: Challenger
          number: 4
        victim:
          name: Victim
          number: 2
        status: complete
    ChallengeResponse:
      allOf:
        - $ref: "#/components/schemas/Challenge"
      example:
        id: fe5d746a-e582-4cec-b2c8-d0c44e9108d1
        challenge: Do the thing
        odds: 20
        challenger:
          name: Challenger
          number: 0
        victim:
          name: Victim
          number: 7
        status: activated
    CompleteChallengeResponse:
      allOf:
        - $ref: "#/components/schemas/Challenge"
      example:
        id: fe5d746a-e582-4cec-b2c8-d0c44e9108d1
        challenge: Do the thing
        odds: 20
        challenger:
          name: Challenger
          number: 7
        victim:
          name: Victim
          number: 7
        status: complete

    # Request Models
    CreateChallengeRequest:
      type: object
      required:
        - challenger
        - victim
        - challenge
      properties:
        challenger:
          type: string
          example: Scott
          minLength: 2
          maxLength: 100
        victim:
          type: string
          example: Michael
          minLength: 2
          maxLength: 100
        challenge:
          type: string
          example: Do the thing
          minLength: 2
    ActivateChallengeRequest:
      type: object
      required:
        - odds
        - number
      properties:
        odds:
          type: number
        number:
          type: number
    CompleteChallengeRequest:
      type: object
      required:
        - number
      properties:
        number:
          type: number

    # Models
    Challenge:
      required:
        - id
        - challenge
        - odds
        - challenger
        - victim
        - status
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: fe5d746a-e582-4cec-b2c8-d0c44e9108d1
        challenge:
          type: string
          example: Do the thing
        odds:
          type: number
        challenger:
          $ref: "#/components/schemas/Person"
        victim:
          $ref: "#/components/schemas/Person"
        status:
          type: string
          enum: [new, activated, complete]
    Person:
      type: object
      properties:
        name:
          type: string
          example: Scott
        number:
          type: number